+++ src/Data_Structure/Queue/Leetcode_Problems/Number_of_Recent_Calls/RecentCounter.cpp
@@ -1,0 +1,11 @@
+class RecentCounter {
+public:
+    queue <int> qu;
+
+
+    int ping(int t) {
+        qu.push(t); //insert ping time in queue
+        while(qu.front() < t-3000) qu.pop(); //remove all pings before 3000ms of current time
+        return qu.size();
+    }
+};
